requirements.txt
aiosqlite==0.21.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.2.1
TgCrypto==1.2.5
typing_extensions==4.15.0
